+++ src/main/java/si/matjazcerkvenik/alertmonitor/web/MultiTagRenderer.java
@@ -40,7 +40,7 @@ public class MultiTagRenderer extends Renderer {
 		
 		//Tags tags = (Tags) component.getAttributes().get("value");
 		// TODO how to get selected provider down here?
-		String dataProviderKey = (String) component.getAttributes().getOrDefault("provider", "/alertmonitor/webhook");
+		String dataProviderKey = (String) component.getAttributes().getOrDefault("provider", ".default");
 		for (Object o:
 			 component.getAttributes().values()) {
 			System.out.println("component attribute: " + o.toString());